requirements.txt
meshtastic==2.7.7
paho-mqtt==2.1.0
requests==2.32.5
slack-sdk
slack_bolt
meshtastic-mqtt-json
cryptography